Adding ticket life cycle logic+test

+++ chickenfarms-app/src/main/java/com/chickenfarms/chickenfarms/repository/UserRepository.java
@@ -1,11 +1,11 @@
 package com.chickenfarms.chickenfarms.repository;
 
-import com.chickenfarms.chickenfarms.model.User;
+import com.chickenfarms.chickenfarms.model.entities.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
 
 }
